+++ packages/at_chops/lib/src/at_chops_impl.dart
@@ -71,7 +71,7 @@ class AtChopsImpl extends AtChops {
     try {
       final decryptionResult = decryptBytes(
           base64Decode(data), encryptionKeyType,
-          encryptionAlgorithm: encryptionAlgorithm, iv: iv);
+          encryptionAlgorithm: encryptionAlgorithm, keyName: keyName, iv: iv);
       final atEncryptionResult = AtEncryptionResult()
         ..atEncryptionMetaData = decryptionResult.atEncryptionMetaData
         ..atEncryptionResultType = AtEncryptionResultType.string;
@@ -125,7 +125,7 @@ class AtChopsImpl extends AtChops {
       final utfEncodedData = utf8.encode(data);
       final encryptionResult = encryptBytes(
           Uint8List.fromList(utfEncodedData), encryptionKeyType,
-          encryptionAlgorithm: encryptionAlgorithm, iv: iv);
+          keyName: keyName, encryptionAlgorithm: encryptionAlgorithm, iv: iv);
       final atEncryptionResult = AtEncryptionResult()
         ..atEncryptionMetaData = encryptionResult.atEncryptionMetaData
         ..atEncryptionResultType = AtEncryptionResultType.string;