+++ packages/at_auth/CHANGELOG.md
@@ -1,2 +1,4 @@
+## 1.0.1
+- feat: Introduce "submitEnrollment" and "manageEnrollment" methods for APKAM
 ## 1.0.0
 - Implemented onboard and authenticate methods.