+++ packages/at_commons/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 4.0.0
+- [Breaking Change] fix: Updated regex for Reserved keys (Internal keys used by the server)
+- fix: Add "put" operation to OTP verb to store semi-permanent pass codes
+## 3.0.58
+- fix: Deprecate encryptedDefaultEncryptedPrivateKey in EnrollParams and introduce encryptedDefaultEncryptedPrivateKey for readability
+- fix: Replace encryptedDefaultEncryptedPrivateKey with encryptedDefaultEncryptionPrivateKey in EnrollVerbBuilder
 ## 3.0.57
 - feat: Introduced TTL(Time to Live) for OTP verb to configure OTP expiry
 ## 3.0.56